+++ lib/graphql_rails/controller/build_controller_action_resolver/controller_action_resolver.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'graphql_rails/controller/request'
+require 'graphql_rails/types/argument_type'
 
 module GraphqlRails
   class Controller
@@ -8,6 +9,8 @@ module GraphqlRails
       # Resolver which includes controller specific methods.
       # Used to simplify resolver build for each controller action
       class ControllerActionResolver < GraphQL::Schema::Resolver
+        argument_class(GraphqlRails::Types::ArgumentType)
+
         def self.controller(controller_class = nil)
           @controller = controller_class if controller_class
           @controller